ensure shortcuts appear
