+++ new-tab-page/new-tab-page-api/src/main/java/com/duckduckgo/newtabpage/api/NewTabPagePlugin.kt
@@ -18,14 +18,14 @@ package com.duckduckgo.newtabpage.api
 
 import android.content.Context
 import android.view.View
-import com.duckduckgo.common.utils.plugins.ActivePluginPoint
+import com.duckduckgo.common.utils.plugins.ActivePlugin
 
 /**
  * This class is used to provide one of the two different version of NewTabPage
  * Legacy -> What existed before https://app.asana.com/0/1174433894299346/1207064372575037
  * New -> Implementation of https://app.asana.com/0/1174433894299346/1207064372575037
  */
-interface NewTabPagePlugin : ActivePluginPoint.ActivePlugin {
+interface NewTabPagePlugin : ActivePlugin {
 
     /** Name of the focused view version */
     val name: String
