+++ new-tab-page/new-tab-page-api/src/main/java/com/duckduckgo/newtabpage/api/NewTabPageSectionPlugin.kt
@@ -18,13 +18,13 @@ package com.duckduckgo.newtabpage.api
 
 import android.content.Context
 import android.view.View
-import com.duckduckgo.common.utils.plugins.ActivePluginPoint
+import com.duckduckgo.common.utils.plugins.ActivePlugin
 
 /**
  * This class is used to provide each of the Sections that build the New Tab Page
  * Implementation of https://app.asana.com/0/1174433894299346/12070643725750
  */
-interface NewTabPageSectionPlugin : ActivePluginPoint.ActivePlugin {
+interface NewTabPageSectionPlugin : ActivePlugin {
 
     /** Name of the focused view version */
     val name: String
