--- new-tab-page/new-tab-page-impl/src/main/java/com/duckduckgo/newtabpage/impl/FocusedPage.kt
@@ -1,38 +1,0 @@
-/*
- * Copyright (c) 2024 DuckDuckGo
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.duckduckgo.newtabpage.impl
-
-import android.content.Context
-import android.view.View
-import com.duckduckgo.anvil.annotations.ContributesActivePlugin
-import com.duckduckgo.di.scopes.ActivityScope
-import com.duckduckgo.newtabpage.api.FocusedViewPlugin
-import com.duckduckgo.newtabpage.api.FocusedViewVersion
-import com.duckduckgo.newtabpage.impl.view.FocusedView
-import javax.inject.Inject
-
-@ContributesActivePlugin(
-    scope = ActivityScope::class,
-    boundType = FocusedViewPlugin::class,
-)
-class FocusedPage @Inject constructor() : FocusedViewPlugin {
-
-    override val name: String = FocusedViewVersion.NEW.name
-    override fun getView(context: Context): View {
-        return FocusedView(context)
-    }
-}
