+++ new-tab-page/new-tab-page-impl/src/main/java/com/duckduckgo/newtabpage/impl/RealNewTabPageSectionProvider.kt
@@ -30,7 +30,7 @@ import kotlinx.coroutines.flow.flow
     scope = AppScope::class,
 )
 class RealNewTabPageSectionProvider @Inject constructor(
-    private val newTabPageSections: ActivePluginPoint<@JvmSuppressWildcards NewTabPageSectionPlugin>,
+    private val newTabPageSections: ActivePluginPoint<NewTabPageSectionPlugin>,
 ) : NewTabPageSectionProvider {
     override fun provideSections(): Flow<List<NewTabPageSectionPlugin>> = flow {
         emit(newTabPageSections.getPlugins().map { it })
