--- new-tab-page/new-tab-page-impl/src/main/res/layout/view_placeholder.xml
@@ -1,31 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-
-
-<!--
-  ~ Copyright (c) 2018 DuckDuckGo
-  ~
-  ~ Licensed under the Apache License, Version 2.0 (the "License");
-  ~ you may not use this file except in compliance with the License.
-  ~ You may obtain a copy of the License at
-  ~
-  ~     http://www.apache.org/licenses/LICENSE-2.0
-  ~
-  ~ Unless required by applicable law or agreed to in writing, software
-  ~ distributed under the License is distributed on an "AS IS" BASIS,
-  ~ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-  ~ See the License for the specific language governing permissions and
-  ~ limitations under the License.
-  -->
-
-<merge xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools">
-
-    <FrameLayout
-        android:layout_width="match_parent"
-        android:background="@color/red20"
-        android:layout_height="200dp"
-        android:orientation="vertical">
-
-    </FrameLayout>
-</merge>
