+++ saved-sites/saved-sites-impl/src/main/java/com/duckduckgo/savedsites/impl/newtab/BookmarksShortcut.kt
@@ -16,44 +16,12 @@
 
 package com.duckduckgo.savedsites.impl.newtab
 
-import android.content.Context
-import android.util.AttributeSet
-import android.widget.LinearLayout
 import com.duckduckgo.anvil.annotations.ContributesActivePlugin
-import com.duckduckgo.anvil.annotations.InjectWith
-import com.duckduckgo.browser.api.ui.BrowserScreens.BookmarksScreenNoParams
-import com.duckduckgo.common.ui.viewbinding.viewBinding
 import com.duckduckgo.di.scopes.AppScope
-import com.duckduckgo.di.scopes.ViewScope
-import com.duckduckgo.mobile.android.databinding.RowNewTabGridItemBinding
-import com.duckduckgo.navigation.api.GlobalActivityStarter
 import com.duckduckgo.newtabpage.api.NewTabPageShortcutPlugin
 import com.duckduckgo.newtabpage.api.NewTabShortcut
 import com.duckduckgo.newtabpage.api.NewTabShortcut.Bookmarks
-import dagger.android.support.AndroidSupportInjection
 import javax.inject.Inject
-
-@InjectWith(ViewScope::class)
-class BookmarksShortcutView @JvmOverloads constructor(
-    context: Context,
-    attrs: AttributeSet? = null,
-    defStyle: Int = 0,
-) : LinearLayout(context, attrs, defStyle) {
-
-    @Inject
-    lateinit var globalActivityStarter: GlobalActivityStarter
-
-    private val binding: RowNewTabGridItemBinding by viewBinding()
-
-    override fun onAttachedToWindow() {
-        AndroidSupportInjection.inject(this)
-        super.onAttachedToWindow()
-
-        binding.root.setOnClickListener {
-            globalActivityStarter.start(this.context, BookmarksScreenNoParams)
-        }
-    }
-}
 
 @ContributesActivePlugin(
     AppScope::class,
