--- saved-sites/saved-sites-impl/src/main/res/layout/view_bookmarks_shortcut.xml
@@ -1,25 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?><!--
-   ~ Copyright (c) 2018 DuckDuckGo
-   ~
-   ~ Licensed under the Apache License, Version 2.0 (the "License");
-   ~ you may not use this file except in compliance with the License.
-   ~ You may obtain a copy of the License at
-   ~
-   ~     http://www.apache.org/licenses/LICENSE-2.0
-   ~
-   ~ Unless required by applicable law or agreed to in writing, software
-   ~ distributed under the License is distributed on an "AS IS" BASIS,
-   ~ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-   ~ See the License for the specific language governing permissions and
-   ~ limitations under the License.
-   -->
-
-
-<com.duckduckgo.common.ui.view.listitem.DaxGridItem
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    app:primaryText="@string/newTabPageShortcutBookmarks"
-    app:leadingIcon="@drawable/ic_bookmarks_open_color_16"
-    app:gridItemType="shortcut"
-    android:layout_width="match_parent"
-    android:layout_height="wrap_content"/>
